'divide_path_crawler_in_different_module
