+++ .gitignore
@@ -1,0 +1,5 @@
+# vscode
+.vscode
+
+# exclude gitkeep
+!.gitkeep
