+++ docs/topics/reader.md
@@ -1,0 +1,2 @@
+## Reader
+This document describes the reader module and format of input data.
